refactor: rename `player:video` module to `player:local`
