+++ app/phone/build.gradle.kts
@@ -103,7 +103,7 @@ dependencies {
     implementation(projects.core)
     implementation(projects.data)
     implementation(projects.player.core)
-    implementation(projects.player.video)
+    implementation(projects.player.local)
     implementation(projects.setup)
     implementation(projects.modes.film)
     implementation(projects.settings)
