+++ app/phone/src/main/java/dev/jdtech/jellyfin/presentation/player/TrackSelectionDialogFragment.kt
@@ -1,4 +1,4 @@
-package dev.jdtech.jellyfin.dialogs
+package dev.jdtech.jellyfin.presentation.player
 
 import android.app.Dialog
 import android.os.Bundle
@@ -8,9 +8,9 @@ import androidx.core.view.WindowInsetsControllerCompat
 import androidx.fragment.app.DialogFragment
 import androidx.media3.common.C
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import dev.jdtech.jellyfin.getTrackNames
-import dev.jdtech.jellyfin.player.video.R
-import dev.jdtech.jellyfin.viewmodels.PlayerViewModel
+import dev.jdtech.jellyfin.player.local.R
+import dev.jdtech.jellyfin.player.local.domain.getTrackNames
+import dev.jdtech.jellyfin.player.local.presentation.PlayerViewModel
 import java.lang.IllegalStateException
 
 class TrackSelectionDialogFragment(
