+++ app/phone/src/main/java/dev/jdtech/jellyfin/utils/PlayerGestureHelper.kt
@@ -27,13 +27,14 @@ import coil3.transform.RoundedCornersTransformation
 import dev.jdtech.jellyfin.PlayerActivity
 import dev.jdtech.jellyfin.core.Constants
 import dev.jdtech.jellyfin.isControlsLocked
-import dev.jdtech.jellyfin.models.PlayerChapter
-import dev.jdtech.jellyfin.models.Trickplay
-import dev.jdtech.jellyfin.mpv.MPVPlayer
+import dev.jdtech.jellyfin.player.core.domain.models.PlayerChapter
+import dev.jdtech.jellyfin.player.core.domain.models.Trickplay
+import dev.jdtech.jellyfin.player.local.mpv.MPVPlayer
 import dev.jdtech.jellyfin.settings.domain.AppPreferences
 import kotlinx.coroutines.Dispatchers
 import timber.log.Timber
 import kotlin.math.abs
+import androidx.core.view.isVisible
 
 class PlayerGestureHelper(
     private val appPreferences: AppPreferences,
@@ -429,21 +430,21 @@ class PlayerGestureHelper(
     private fun releaseAction(event: MotionEvent) {
         if (event.action == MotionEvent.ACTION_UP || event.action == MotionEvent.ACTION_CANCEL) {
             activity.binding.gestureVolumeLayout.apply {
-                if (visibility == View.VISIBLE) {
+                if (isVisible) {
                     removeCallbacks(hideGestureVolumeIndicatorOverlayAction)
                     postDelayed(hideGestureVolumeIndicatorOverlayAction, 1000)
                     swipeGestureVolumeOpen = false
                 }
             }
             activity.binding.gestureBrightnessLayout.apply {
-                if (visibility == View.VISIBLE) {
+                if (isVisible) {
                     removeCallbacks(hideGestureBrightnessIndicatorOverlayAction)
                     postDelayed(hideGestureBrightnessIndicatorOverlayAction, 1000)
                     swipeGestureBrightnessOpen = false
                 }
             }
             activity.binding.progressScrubberLayout.apply {
-                if (visibility == View.VISIBLE) {
+                if (isVisible) {
                     if (swipeGestureValueTrackerProgress > -1) {
                         playerView.player?.seekTo(swipeGestureValueTrackerProgress)
                     }
