+++ player/local/src/main/res/values-bg/strings.xml
