+++ player/local/src/main/res/values-cs-rCZ/strings.xml
