+++ player/local/src/main/res/values-es-rMX/strings.xml
