+++ player/local/src/main/res/values-es/strings.xml
