+++ player/local/src/main/res/values-fr/strings.xml
