+++ player/local/src/main/res/values-iw/strings.xml
