+++ player/local/src/main/res/values-ko/strings.xml
