+++ player/local/src/main/res/values-pl/strings.xml
