+++ player/local/src/main/res/values-pt-rBR/strings.xml
