+++ player/local/src/main/res/values-pt/strings.xml
