+++ player/local/src/main/res/values-ru/strings.xml
