+++ player/local/src/main/res/values-sk/strings.xml
