+++ player/local/src/main/res/values-sl/strings.xml
