+++ player/local/src/main/res/values-sv/strings.xml
