+++ player/local/src/main/res/values-tr/strings.xml
