+++ player/local/src/main/res/values-zh-rCN/strings.xml
