+++ player/local/src/main/res/values/strings.xml
