+++ settings.gradle.kts
@@ -7,7 +7,7 @@ include(":app:tv")
 include(":core")
 include(":data")
 include(":player:core")
-include(":player:video")
+include(":player:local")
 include(":setup")
 include(":modes:film")
 include(":settings")
